Update integrator.xml
DITA OT 2.5+ replaced the "preprocess" stage with "preprocess2" for monolithic output formats. "preprocess2" writes unique filenames to the temp folder, which breaks the pubmeta filter (pubmeta looks for the original map filename for additional processing). Submitting this as a temporary fix to allow TC spec builds with newer versions of the OT.